pyproject.toml
[tool.poetry]
name = "resume"
version = "0.1.0"
description = ""
authors = ["GuillaumeO <7831777+GuillaumeOj@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
dj-database-url = "^0.5.0"
whitenoise = "^5.1.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.5"
flake8-django = "^1.1.1"
isort = "^5.2.2"
docformatter = "^1.3.1"

[tool.black]
line-length = 90

[tool.isort]
force_single_line = true
lines_after_imports = 2
force_sort_within_sections = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
